Add notify replication failed for TRANSPORT_ERROR & REPOSITORY_MISSING
We send replication started notification for each attempt to push.
If the push fails with a reason TRANSPORT_ERROR or REPOSITORY_MISSING,
it will be retried until it succeed (endless loop in case
TRANSPORT_ERROR that occurs forever).

For each new re-try we will send another replication started
notification for that branch, but we will miss the notification
about failed attempt completed. This patch adds the missing
notification.

This fixes notification on push to replica that is down.

Change-Id: I6d5d06fbc081df2d9a32973af73f6a2136d51731
Signed-off-by: Eryk Szymanski <eryksz@gmail.com>