Update patch set 1
Patch Set 1:

> I forgot about these pending changes and they didn't get included
 > in 2.13.9 :(

No worries, that's probably my fault as I did not manage to verify this in time.

We actually *did* try to verify the fix, but at least for our use-case it did not help as expected. Here's what we did:

- Configure two replica servers and make sure replication goes through to both of them

- Reload the patched replication plugin

- Remove ssh credentials on the second replica side, so that authentication from the master would fail and this would prevent the replication from succeeding

- Re-replicate all the repositories to both replicas

- ssh -p 29418 qa-gerrit.it.here.com replication start –all

Result: Gerrit queue congestion is still being observed. All the push events to the failing replica stay in the replication queue and get rescheduled forever. We were hoping for this patch to fix exactly that issue.

Patch-set: 1
Reviewer: Sebastian Schuberth <1015513@173816e5-2b9a-37c3-8a2e-48639d4f1153>