Check the result value of remote ssh commands
When performing remote ssh commands in ReplicationQueue, the result
value of the remote commands was ignored. This resulted in logs stating
that everything went fine, while in reality something went wrong.

This will check the exit value of the remote command, and print to the
error log in case of a non zero value.

Change-Id: Icf0c4adedfa2d998585606c5ef68c58ea7bdc46c