Throw exception on failed remote command
There are (almost) no checks of the exit value of remote ssh commands.
This will cause the replication log to contain successful operations,
when they in reality failed.

This commit will throw a subclass of IOException when the ssh commands
returns a non zero value. Since the IOException is already handled, we
will get proper logging of these errors "for free".

The only place where we previously handled the return value was in
delete project, here we will explicitly check the exception type, and
handle the error value as before.

Change-Id: I29bd8c92141fea9dadbfd7e3896395376eb509eb