Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > > Patch Set 2:
>  > >
>  > > > Move this to the stable branch?
>  > >
>  > > +1 for 2.14
>  > 
>  > This change doesn't cherry-pick cleanly onto stable-2.14 because
>  > the GerritSshApi class doesn't exist there.
> 
> I looked and the problem fixed by this commit exists on stable-2.14, exit codes are not handled. The code to execute ssh command is in ReplicationQueue on stable-2.14, SshHelper did not exist yet.
> 
> IIRC, SshHelper was added in 2.15 so cherry-pick on stable-2.15 should be easier.

Yes, it cherry-picks cleanly on stable-2.15.

If we want this for stable-2.14 it will have to be reworked.  For now I think we can move this to stable-2.15, and then later do a separate fix targetted to stable-2.14.

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>