Fix creating missing repository
When replicating to a destination where the repository does not exist,
updating the HEAD reference failed because the passed reference name was
not absolute. When this happened, an unchecked exception was triggered
but never showed in the logs. In fact, the replication logs were showing
only the message announcing the replication started, but no additional
information about the success or failure of the operation was displayed.

Avoid this by resolving symbolic references and checking the reference
is absolute before trying to update the HEAD in the new repository.

Change-Id: I18295a37a04413ba64bf7e9ee31640ee23c4c1e0