Fix creating missing repository
When replicating to a local destination, if the repository does not
exist, updating the HEAD ref fails because the passed ref name is not
absolute. When this happens, an unchecked exception is triggered but
never showed. In the logs, only the message announcing the replication
started is shown but no more information (creation succeeded or failed)
is displayed.

Checking if the reference is absolute before trying to update the
reference to HEAD avoids this error and shows the repository was
created.

Change-Id: I18295a37a04413ba64bf7e9ee31640ee23c4c1e0