Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Is this fixing the same issue that was fixed in [1]?
> > 
> > That change is only on stable-2.15 and master.
> > 
> > [1] https://gerrit-review.googlesource.com/c/plugins/replication/+/107350
> 
> Indeed, seems to be the same. Sorry, didn't notice this one.

The difference is, the way this change does it means the creation succeeds, the HEAD is pointing to the right reference and the success of the operation is shown in the logs, which was not the case before. However I guess we can ditch this one and wait for the upcoming 2.15 to fix this problem?

Patch-set: 2
Reviewer: Hector Oswaldo Caballero <1019328@173816e5-2b9a-37c3-8a2e-48639d4f1153>