Update patch set 2
Patch Set 2:

> Patch Set 2: Verified+1 Code-Review+1
> 
> It is passing though our CI however we don't have tests for local replication.

In fact, this should fix both local and remote replications: the problem was that when calling createProject (PushOne L.403) we were passing the name of the reference. If the reference is a Symbolic one, let's say, HEAD, the name of the reference is "HEAD", but in fact what ReplicationQueue#createLocally (L. 264) and ReplicationQueue.createRemote (L.275) are expecting is the target, i.e., where the reference is pointing to, like 'refs/heads/master', for example. When we handle the symbolic references, the right parameter is passed and both create repository methods succeed creating the repo (as it was before) and setting the HEAD to the right reference.

The change done in 2.15 that David was referring to seems to accept the fact that the HEAD was not set correctly in the first place and generates a new event with the right reference so that it can be set afterwards, IIUC.

My opinion is that this change is a bit cleaner as it is fixing the original issue, but may be I'm just not understanding correctly the 2.15's change. If my understanding is right, I'd suggest merging this change and reverting the existing change in 2.15 before merging this branch up.

Patch-set: 2
Reviewer: Hector Oswaldo Caballero <1019328@173816e5-2b9a-37c3-8a2e-48639d4f1153>