Remove #isHidden: adjust the "ACCESS" permission check
This commit adjusts the "ACCESS" permission checks in this plugin to
check whether the project state is readable:

* If it's readable, then the "ACCESS" permission will be checked.
* If it's not readable, the project is only accessable by the project
owners. In this case, "READ_CONFIG" will be checked, which is only
holded by the project owners.

This commit should be a no-op and it's an essential step for removing
ProjectControl#isHidden, which is a non-permission check and thus
should not be called from the permission backend.

Change-Id: I68e41cfba6d7c6e31aeae75652755a3f6b53c3dd