Don't loose ref-updated events when on plugin restart
When a ref-updated event is received, persist the event in the directory
defined by the replication.eventsDirectory. When the udpated ref is
replicated deleted the persisted event.

If replication queue is non-empty and plugin gets stopped, ref updates
will not be replicated and, therefore, the persisted events will not get
deleted. When the plugin starts it will schedule replication for all
persisted events and delete them.

A main intention of this change is to avoid the need for the
replicateOnStartup=true setting which schedules replication of all refs
for all projects and typically creates a humongous replication queue on
every plugin restart.

Change-Id: Ieacd084fabe703333241ffda11c8b6c78cced37a