Delete event file only after replication completed for all destinations
The event file is created only once even if the ref is replicated to
more than one destination, but it was deleting after replication is
completed to each destination.

This was wrong, as only after the first destination completes the file
was successfully deleted. And upon subsequent destination completion,
the deletion was failing with NoSuchFileException. On top of the error
getting logged, this is also wrong as the plugin could be reloaded
before replication is completed on all destinations. And if one of them
completed, then replication for that ref would not have been
rescheduled.

Change-Id: I07b3b88e42382330da409dd9e156498c0bf434a9