Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Nice to see these long-standing bugs to be addressed. One question, though, couldn't (shouldn't) these be fixed in a stable version?

Agreed, it would be good to fix these in a stable version.  However, I would prefer not to put them in stable-2.14 to avoid risk of regression (and since we're already on the 16th patch release). Therefore, if we do put these fixes to a stable branch I would prefer stable-2.15 or stable-2.16.

Patch-set: 1
CC: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>