Execute all admin operations from the replication queue thread pool
Before this change, the delete-project and update-head operations were
executed from the calling Gerrit core thread while the create-project
was executed in the context of the PushOne operation which is run from
the replication queue.

Tasks executed from the replication queue are created from an additional
child injector which provides more bindings, among others the
RemoteConfig. Having access to RemoteConfig from an admin task may be
necessary in order to retrieve HTTP credentials, like the PushOne task
does. This will be necessary when we implement admin operations over
the HTTP/REST protocol.

Change-Id: Iae608aa98474d6405b26bbae2508183c2d0cb212