ReplicationFileBasedConfig: Catch exception when creating destination
DestinationFactory gets an intance of GroupIncludeCache injected, which
references the group index.  When the plugin is loaded at site startup,
the group index is not available due to the index module being loaded
after the plugins module per change I87da6ac63. As a result, the plugin
fails to load and a stack trace is logged.

This only occurs when a destination configuration has an authGroup
setting, which causes the group include cache to be used.

If the plugin is later loaded again after the server has started, it is
successful. However this requires the administrator to bump the plugin
jar file's timestamp (otherwise Gerrit won't reload it because it is
unchanged since last being marked as 'broken').

Catch the exception, log an error, and continue with loading remaining
destinations. This will allow the plugin to load successfully, and be
reloaded without having to touch the jar file timestamp

Bug: Issue 10082
Change-Id: I7559faa98e72621f5cf0ba3081f961f6ebe11965