Update patch set 4
Patch Set 4:

I'm not sure it's feasible to get a proper fix for this in short time.  Can we go ahead with this as a temporary workaround for 2.16.1, and then revert it later when we have the proper fix available?

Patch-set: 4
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>