Fix auto-replay of stored events during config reload
Use a proper configuration reload listener to allow ReplicationQueue
to be aware of what's going on and perform a proper stop/start sequence
instead of relying on a low-level WorkQueue access.

The configuration reload behaves now as a soft-restart so that
stored events parked on the filesystem won't be missed any more.

Bug: Issue 10260
Change-Id: Ia34b460221a2e0512a4ffba66d1f08c201402917