Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I think the parent change can be squashed into this.  On the parent change I was going to comment that the log should indicate that not all refs are mentioned,

At least this part could be squashed (indicating that not all refs are mentioned).
When this is done, the two changes still make sense: the first one introduces a limit
and the second one makes the limit configurable.

> but I see it's done in this change.
> 
> An alternative to making this configurable would be to never log all the refs at info level, but rather just say something like "replicating NNN refs...", and move logging of all the refs to a separate log statement at debug level.  WDYT?

Remember that the replication plugin will do a PushAll command at the start (configurable).
The problem is that there are repos with >500K refs and it is just generating a lot of IO
to write such logs, probably also affecting the overall performance. And if this replication
needs to be rescheduled the whole list is being logged again in its full size.

Changing the logging level doesn't help much. Changing the logging level may also break
some tools which monitor the replication_log.

Patch-set: 2
Reviewer: Gerrit User 1003873 <1003873@173816e5-2b9a-37c3-8a2e-48639d4f1153>