Introduce DynamicSet of ReplicationStateListener
The ReplicationStateListener interface is used at the moment for the
sole purpose of logging the replication activity, without any other
implementations. However, the reason why it was introduced was to allow
extra listeners that can take actions on events.

Prefer the DynamicSet binding mechanism so that can allow other extra
listener to the replication events, which is eventually the purpose of
introducing the listener in the first place.

Change-Id: I551d500b95192cf26f12548260b1c126f240b814