Separate Config and Destination and allow more flexible reload
Parsing the configuration and creating the JGit destinations are two
different concerns: refactor the current configuration monolith into
two different classes, introducing the DestinationsCollections class.

Decouple the DestinationsCollection from the knowledge of the reloadable
replication.config.  That allows to get rid of the code-smell of
reloadIfNeeded in ReplicationConfig by having the configuration to call
the listener on-demand and whenever is needed.

Additionally, ReplicationConfig can be reloadable by implementing an ETag that
identifies the version of the file.

Change-Id: Ia1c01de0d5edf5c9976d8da8ee5b255247a1a7a8