Introduce ReplicationConfigListener on loading
Decouple the DestinationsCollection from the knowledge of the reloadable
replication.config.

Allows to get rid of the code-smell of reloadIfNeeded in
ReplicationConfig by having the configuration to call the listener
on-demand and whenever is needed.

Any ReplicationConfig can be reloadable by implementing an ETag that
identifies the version of the file.

Change-Id: I9c53b68a9249c65a3077a4b5fc498bd426f230ec