Use GitRepositoryManager openTransport to return a JGit Transport
The purpose of this change is to remove the direct use of the JGit Transport.open()
and shift this responsibility to be managed by Gerrit's GitRepositoryManager class.

Moving this logic to the GitRepositoryManager class provides also the flexibility
of injecting potential validation checks whenever the replication plugin pushes new
references to the remote server.

Change-Id: I447b3adc881ac38606fb2bbe28d4901dacfa44c5