Extract destinations logic into a new class
Move the logic for creating and querying the destinations to a new
specific class called DestinationsCollection.

Because of not being linked to any specific implementation of the
remote, enable other implementations of the ReplicationConfig
from other plugins (e.g. pull-replication) and also consuming
the same replication.config for other purposes.

Feature: Issue 11425
Change-Id: I253a6d94d58efb30ece8624eb0e757193b09b8c1