ReplicationQueue: set as running only when events replay is finished
The replication queue is responsible for re-triggering all the past
replay events that have not been dispateched yet upon startup.

The queue is ready to be used and serve properly all incoming requests
and background configuration reloads *only when* the full replay is
completed. Move the running flag only after the replay is finished
and allow other components to know its running status.

The configuration auto-reload mechanism can then query the queue
status and not reloading anything until the queue is fully running
and ready to serve all events.

Change-Id: I4aad49f3f854b2675da0631d368ff4725bf4e7a0