ReplicationQueue: set as running only when events replay is finished
Fix a problem where the configuration was reloaded during the
event replay was in-progress. The result was the processing of
events with obsolete Destination objects in memory and triggering
NPE on the dispatching of replication events.

java.lang.NullPointerException
 at Destination.schedule(Destination.java:367)
 at Destination.schedule(Destination.java:326)
 at ReplicationQueue.onGitReferenceUpdated(ReplicationQueue.java:124)
 at ReplicationQueue.firePendingEvents(ReplicationQueue.java:134)
 at ReplicationQueue.start(ReplicationQueue.java:72)

The replication queue is responsible for re-triggering all the past
replay events that have not been dispatched yet upon startup.
Until the re-trigger is complete, the Gerrit start-up sequence is
on hold because the dispatch is done synchronously to the plugin
startup sequence.

The queue is ready to be used and serve properly all incoming requests
and background configuration reloads *only when* the full replay is
completed. Allowing extra events to come through would have some
unexpected consequences because Gerrit isn't ready yet until all
the plugins have been started and the startup of the queue is only
one phase of that startup sequence.

Move the running flag only after the replay is finished
and allow other components to know its running status. Doing so
allows the configuration auto-reload mechanism to query the queue
status and not reloading anything until the queue is fully running
and ready to serve all events.

Bug: Issue 11055
Change-Id: I4aad49f3f854b2675da0631d368ff4725bf4e7a0