Store replication tasks instead of ref-update events
Fix the granularity of persistent events on disk by storing
replication tasks rather than ref-update events.

Ref-updates are the initial cause of the replication tasks,
however it is tricky to perform reference counting on the associated
tasks reliabily, due to all the different conditions of where a
replication tasks can be in the replication queue.

Fixes a recurring issue where persisted ref-update events were missing or
left orphan on the filesystem due to the different granularity of the
incoming ref-update events and the corresponding replication tasks queued
and executed.

The responsibility of persisting the events is moved from the
ReplicationState object to the Destination.

The stored objects are replication tasks and not anymore ref-update
events.

Bug: Issue 11172
Change-Id: I02d8cda0a124e8e3d2b9bb01b7d44f98ba717fcd