Fix granularity of persistent replication events
Align the granularity of persistent events on disk with the replication
events processed by the Destination.

Fixes a recurring issue where persistent events were missing or
left orphan on the filesystem due to the different granularity of the
incoming ref-update events and the corresponding git push operation queued
and executed.

The responsibility of persisting the events is moved from the
ReplicationState object to the Destination.

Bug: Issue 11172
Change-Id: I02d8cda0a124e8e3d2b9bb01b7d44f98ba717fcd