Do not reload config when queue is not ready
Do not stop/start the replication queue upon a new configuration
added when the queue is either not running yet or is replaying
incoming events.

Restarting a partially initialized queue would have unexpected
consequences as the objects that are needed to schedule and
serve the requests could be still null and generate random NPEs
all around.

The replay also is a very delicate phase because it is responsible
for the replay of the past missed replication events that should
be re-triggered with highest priority possible. All new events
and new configuration can wait to be processed once all the
past events have been replayed and have completed.

Bug: Issue 11055
Change-Id: Ib154532ec32a50f4cb9ca5553c1f77fa931ae2ac