Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can this be done on one of the earlier stable branches?  Maybe stable-2.15?

Yes, as soon as my chain of fixes for the Issue 11172 is verified and agreed.
The combination of all those bugs created the above issue and, for making sure that the problem is fully resolved, I need to keep them in line on stable-3.0 for now.

Patch-set: 1
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>