Option to drain the event queue before shutdown
The node can be down for a while and the replication events could be
missing for all that time.
When the node is coming up again, the SHA1 should be then obsolete
and then effectively lost.

Allow the Gerrit administrator to avoid delaying replication events
by automatically making the shutdown of the replication queue more
graceful and wait for all in-flight replications tasks to end before
shutting down the system.

Bug: Issue 11145
Change-Id: I9d823cd64176b4987d7ba71d9a46e717d0f52b93