Update patch set 3
Patch Set 3: Code-Review+1

> Patch Set 3: Code-Review-1
> 
> This is recording the stats for both destination and destination/project using the same metric.  There should be a separate metric introduced for the destination/project stats.

I believe the value of the metrics is the same (timing of the replication operation), however, the granularity of the project was lost. Would you say that having a second recording of the same metric but with the project, granularity would impact the value of the existing one?

Patch-set: 3
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1