Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review+1
> > 
> > > Patch Set 3: Code-Review-1
> > > 
> > > This is recording the stats for both destination and destination/project using the same metric.  There should be a separate metric introduced for the destination/project stats.
> > 
> > I believe the value of the metrics is the same (timing of the replication operation), however, the granularity of the project was lost. Would you say that having a second recording of the same metric but with the project, granularity would impact the value of the existing one?
> 
> With this implementation the same metric is recorded twice; for remote and remote/project. I'm not 100% sure what affect this will have for example when looking at visualizations of that metric.

@Tony can you please verify?

The problem here is that, in theory, you should have one metric per project, that would make the payload literally explode (with all the associated percentiles).

Agreed to double-check, to make sure to not have any impact on the existing metrics.

P.S. It should be moved also to v3.0, as this is a new feature and shouldn't target an existing stable branch unless it is a bug-fix.

Patch-set: 3
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>