AdminApi: Reintroduce return value of methods
Prior to change Ie760bf3e1 all the methods of GerritSshApi returned
a boolean value. After that change, they all return void.

Removing the return value breaks classes that extend GerritSshApi and
expect to be able to invoke the superclass implementation and get the
status.

In change I6566f8671 the return value of createProject was changed
back to boolean, but the others were not.

Change the remaining methods back to return a boolean.

Change-Id: If4a9227f70bfcd85965e1febd45f4a75cc2505bd
Signed-off-by: David Pursehouse <dpursehouse@collab.net>