Update patch set 1
Patch Set 1:

This can be a bit dangerous because it triggers activity on the replication queue that isn't a consequence of a ref-update event.

The overall replication state is super-flaky at the moment and poorly tested: I am afraid that it could compromise even further the fragile stability of the plugin in v2.16.

I would like to get another couple of eyes (CC Saša) on it.

Patch-set: 1
Reviewer: Gerrit User 1003873 <1003873@173816e5-2b9a-37c3-8a2e-48639d4f1153>
CC: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>