Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > This can be a bit dangerous because it triggers activity on the replication queue that isn't a consequence of a ref-update event.
> > > 
> > 
> > This is only a risk when the plugin is extended and the extending implementation makes use of the method.  For the vanilla version, this is a no-op change.
> 
> Yes, I know. However, as you know, once you open the pandorra-box for people to use a feature, then they start using it and, if doesn't work or break anything else, they'll start complaining :-(
> 
> Did you guys perform testing on various scenarios of the impacts of the change?
> Do you believe you could add some coverage to demonstrate that, when used, doesn't break other stuff?

We have tests for replication, but it's the teamforge replication that's based upon this.  We have this patch internally for some years on 2.14 and 2.15, and it's used in production.

I can't test it yet for 2.16 because I still need to also migrate a lot of other stuff to 2.16.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>